+++ core-utils/src/commonTest/kotlin/com/pambrose/util/StringExtensionTests.kt
@@ -50,6 +50,8 @@ import com.pambrose.common.util.toSingleQuoted
 import com.pambrose.common.util.trimEnds
 import io.kotest.core.spec.style.StringSpec
 import io.kotest.matchers.shouldBe
+import io.kotest.matchers.string.shouldMatch
+import io.kotest.matchers.string.shouldNotMatch
 
 class StringExtensionTests : StringSpec() {
   init {
@@ -279,6 +281,28 @@ class StringExtensionTests : StringSpec() {
       "Test.java".contains("T?s?.*a".asRegex()) shouldBe true
       "Test.java".contains("t?s?.*a".asRegex()) shouldBe false
       "Test.java".contains("t?s?.*a".asRegex(true)) shouldBe true
+    }
+
+    "pattern match treats regex metacharacters in the glob literally" {
+      "file(1).txt" shouldMatch "file(1).txt".asRegex()
+      "file1.txt" shouldNotMatch "file(1).txt".asRegex()
+      "a+b" shouldMatch "a+b".asRegex()
+      "aab" shouldNotMatch "a+b".asRegex()
+      "[abc" shouldMatch "[abc".asRegex()
+      "a\$b^c{1}|d\\e" shouldMatch "a\$b^c{1}|d\\e".asRegex()
+      "a__SINGLE__DOT__b".toPattern shouldBe "^a__SINGLE__DOT__b$"
+    }
+
+    "Long length is exact at every power-of-ten boundary" {
+      var power = 1L
+      for (digits in 1..18) {
+        power *= 10
+        (power - 1).length shouldBe digits
+        power.length shouldBe digits + 1
+        (-(power - 1)).length shouldBe digits
+      }
+      Long.MAX_VALUE.length shouldBe 19
+      Long.MIN_VALUE.length shouldBe 19
     }
   }
 }